[ALPS08003864] ACK: Merge android12-5.10-2023-03 R1 into alps-mp-s0.mp1
[Target]:
    62f5fae173d8afb19749a1b78d8f11fb23d92164
    UPSTREAM: ext4: refuse to create ea block when umounted

[Source:] kernel/common android12-5.10-2023-03
[From:] 8d44cb1a9e42 ANDROID: dma-buf: heaps: Dont lock unused dmabuf_page_pool mutex
[  To:] 62f5fae173d8 UPSTREAM: ext4: refuse to create ea block when umounted

MTK-Commit-Id: ca0e8ed876a24890754fd9662c45c77c8ed94343

Feature: Kernel SI Operation
CR-Id: ALPS08003864

+++ arch/x86/kvm/vmx/vmx.c
@@ -1431,8 +1431,10 @@ void vmx_vcpu_load_vmcs(struct kvm_vcpu *vcpu, int cpu,
 
 		/*
 		 * No indirect branch prediction barrier needed when switching
-		 * the active VMCS within a guest, e.g. on nested VM-Enter.
-		 * The L1 VMM can protect itself with retpolines, IBPB or IBRS.
+		 * the active VMCS within a vCPU, unless IBRS is advertised to
+		 * the vCPU.  To minimize the number of IBPBs executed, KVM
+		 * performs IBPB on nested VM-Exit (a single nested transition
+		 * may switch the active VMCS multiple times).
 		 */
 		if (!buddy || WARN_ON_ONCE(buddy->vmcs != prev))
 			indirect_branch_prediction_barrier();
@@ -1737,7 +1739,17 @@ static void vmx_queue_exception(struct kvm_vcpu *vcpu)
 	kvm_deliver_exception_payload(vcpu);
 
 	if (has_error_code) {
-		vmcs_write32(VM_ENTRY_EXCEPTION_ERROR_CODE, error_code);
+		/*
+		 * Despite the error code being architecturally defined as 32
+		 * bits, and the VMCS field being 32 bits, Intel CPUs and thus
+		 * VMX don't actually supporting setting bits 31:16.  Hardware
+		 * will (should) never provide a bogus error code, but AMD CPUs
+		 * do generate error codes with bits 31:16 set, and so KVM's
+		 * ABI lets userspace shove in arbitrary 32-bit values.  Drop
+		 * the upper bits to avoid VM-Fail, losing information that
+		 * does't really exist is preferable to killing the VM.
+		 */
+		vmcs_write32(VM_ENTRY_EXCEPTION_ERROR_CODE, (u16)error_code);
 		intr_info |= INTR_INFO_DELIVER_CODE_MASK;
 	}
